Initial Commit
This is a game of blackjack that I coded yesterday. My girlfriend is taking a coding class and they had an assignment like this, but they're doing it in python, and I figured it could be a fun activity for me to practice my c# skills. I have been quite delighted to find that c# is very similar to Java, which is the last language I worked in, so it was a bit easier to pick up c# than I was expecting.