created + printed array using foreach loop + created function Convert.toInt()
For this exercise, I thought it might be cool to let the user decide how long the array is and then what each of the items in the array will be. To prevent the program from crashing when the user enters a non-integer value in response to being asked how long they want the array, I created a Convert class with a toInt() function. Convert.toInt() has one parameter (a string variable that is what's being converted to an int), and it returns a Convert object. The convert object has two variables attached to it - one bool, and one int. The bool value returns false if the string cannot be successfully converted into an int, and returns true if the string can be successfully converted into an int. The int value is the converted value of the string