scoreboard no longer resets between scenes
The ArrayObj object Scoreboard was resetting between scenes. It would always do this unless you were looking at it in the inspector as you switched scenes. I looked some stuff up and I found out that apparently if a scene doesn't ever reference an asset, it'll reset when that scene is loaded. To solve this I just made a script that references Scoreboard a single time in the Start() function. I can attach the script to an empty object in any scene that doesn't already reference Scoreboard. I playtested to see if this solution works, and as far as I can tell, it works just fine, and since it only references Scoreboard once at the start of the scene loading, it shouldn't use hardly any data and shouldn't really make the performance any worse